refactor: rename and restructure InputIcon component
